requirements.txt
python-dotenv
pandas
scikit-learn
joblib

